MessageInterface Rev 반영
