+++ src/I2CInterface/I2CInterface.cpp
@@ -1,6 +1,5 @@
 #include <Arduino.h>
 #include "I2CInterface.h"
-#include "Endian.h"
 
 uint8_t I2CInterface::i2cReadBuffer[I2C_INTERFACE_BUFFER_SIZE] = {
     0,
@@ -13,376 +12,7 @@ uint8_t I2CInterface::i2cWriteBuffer[I2C_INTERFACE_BUFFER_SIZE] = {
 size_t I2CInterface::i2cWriteBufferIdx = 0;
 size_t I2CInterface::i2cWriteBufferSize = 0;
 
-I2CInterface::CallBack::CallBackRequestGet I2CInterface::CallBack::_callBackRequestGet = NULL;
-I2CInterface::CallBack::CallBackResponseGet I2CInterface::CallBack::_callBackResponseGet = NULL;
-I2CInterface::CallBack::CallBackRequestSet I2CInterface::CallBack::_callBackRequestSet = NULL;
-I2CInterface::CallBack::CallBackSubscribe I2CInterface::CallBack::_callBackSubscribe = NULL;
-I2CInterface::CallBack::CallBackUnsubscribe I2CInterface::CallBack::_callBackUnsubscribe = NULL;
-I2CInterface::CallBack::CallBackNotification I2CInterface::CallBack::_callBackNotification = NULL;
-I2CInterface::CallBack::CallBackAcknowledge I2CInterface::CallBack::_callBackAcknowledge = NULL;
-
 int I2CInterface::I2C_Slave::commandType = 0;
-
-uint8_t I2CInterface::getCheckSum(const uint8_t *str, size_t strSize)
-{
-    uint8_t sum = 0;
-    for (size_t idx = 0; idx < strSize; idx++)
-    {
-        sum = sum + str[idx];
-    }
-
-    uint8_t checkSum = (sum ^ 0xFF);
-
-    // cout << "CheckSum : " << (int)checkSum << endl;
-    return checkSum;
-}
-
-bool I2CInterface::verityCheckSum(const uint8_t *str, size_t strSize)
-{
-    uint8_t sum = 0;
-    for (size_t idx = 0; idx < strSize; idx++)
-    {
-        sum = sum + str[idx];
-    }
-
-    // cout << "Verity : " << (sum == 0) << endl;
-    return (sum == 0);
-}
-
-void I2CInterface::writeEndian(ResponseGetData *data)
-{
-    data->value = htonl(data->value);
-}
-
-void I2CInterface::readEndian(ResponseGetData *data)
-{
-    data->value = ntohl(data->value);
-}
-
-void I2CInterface::CallBack::registerRequestGetCallBack(const CallBackRequestGet func)
-{
-    _callBackRequestGet = func;
-}
-
-void I2CInterface::CallBack::invokeRequestGetCallBack(uint32_t seqId, const RequestGetData *dataArr, size_t arrSize)
-{
-    if (_callBackRequestGet == NULL)
-        return;
-    _callBackRequestGet(seqId, dataArr, arrSize);
-}
-
-void I2CInterface::CallBack::registerResponseGetCallBack(const CallBackResponseGet func)
-{
-    _callBackResponseGet = func;
-}
-
-void I2CInterface::CallBack::invokeResponseGetCallBack(uint32_t seqId, const ResponseGetData *dataArr, size_t arrSize)
-{
-    if (_callBackResponseGet == NULL)
-        return;
-
-    for (size_t arrIdx = 0; arrIdx < arrSize; arrIdx++)
-    {
-        I2CInterface::readEndian((ResponseGetData *)(dataArr + arrIdx));
-    }
-
-    _callBackResponseGet(seqId, dataArr, arrSize);
-}
-
-void I2CInterface::CallBack::registerNotificationCallBack(const CallBackNotification func)
-{
-    _callBackNotification = func;
-}
-
-void I2CInterface::CallBack::invokeNotificationCallBack(uint32_t seqId, const NotificationData *dataArr, size_t arrSize)
-{
-    if (_callBackNotification == NULL)
-        return;
-
-    for (size_t arrIdx = 0; arrIdx < arrSize; arrIdx++)
-    {
-        I2CInterface::readEndian((ResponseGetData *)(dataArr + arrIdx));
-    }
-
-    _callBackNotification(seqId, dataArr, arrSize);
-}
-
-void I2CInterface::CallBack::registerSubscribeCallBack(const CallBackSubscribe func)
-{
-    _callBackSubscribe = func;
-}
-
-void I2CInterface::CallBack::invokeSubscribeCallBack(uint32_t seqId, const SubscribeData *dataArr, size_t arrSize)
-{
-    if (_callBackSubscribe == NULL)
-        return;
-    _callBackSubscribe(seqId, dataArr, arrSize);
-}
-
-void I2CInterface::CallBack::registerUnsubscribeCallBack(const CallBackUnsubscribe func)
-{
-    _callBackUnsubscribe = func;
-}
-
-void I2CInterface::CallBack::invokeUnsubscribeCallBack(uint32_t seqId, const UnsubscribeData *dataArr, size_t arrSize)
-{
-    if (_callBackUnsubscribe == NULL)
-        return;
-    _callBackUnsubscribe(seqId, dataArr, arrSize);
-}
-
-void I2CInterface::CallBack::registerRequestSetCallBack(const CallBackRequestSet func)
-{
-    _callBackRequestSet = func;
-}
-
-void I2CInterface::CallBack::invokeRequestSetCallBack(uint32_t seqId, const RequestSetData *dataArr, size_t arrSize)
-{
-    if (_callBackRequestSet == NULL)
-        return;
-
-    for (size_t arrIdx = 0; arrIdx < arrSize; arrIdx++)
-    {
-        I2CInterface::readEndian((ResponseGetData *)(dataArr + arrIdx));
-    }
-
-    _callBackRequestSet(seqId, dataArr, arrSize);
-}
-
-void I2CInterface::CallBack::registerAcknowledgeCallBack(const CallBackAcknowledge func)
-{
-    _callBackAcknowledge = func;
-}
-
-void I2CInterface::CallBack::invokeAcknowledgeCallBack(uint32_t seqId, unsigned char msgId)
-{
-    if (_callBackAcknowledge == NULL)
-        return;
-    _callBackAcknowledge(seqId, msgId);
-}
-
-void I2CInterface::MessageReceiver::listen()
-{
-    // Serial.printf("[%s:%d] Buffer size : %d\n", __FUNCTION__, __LINE__, i2cReadBufferIdx);
-    // if(i2cReadBufferIdx == 0) return;
-    // for (int i = 0; i < i2cReadBufferIdx; i++)
-    // {
-    //     Serial.printf("%d(%d)", i, (int)i2cReadBuffer[i]);
-    //     Serial.print(' ');
-    // }
-    // Serial.println();
-
-    bool begin = false;
-    if (i2cReadBufferIdx >= 7)
-    {
-        if (memcmp(i2cReadBuffer, "<BEGIN>", 7) == 0)
-        {
-            begin = true;
-        }
-    }
-
-    bool end = false;
-    if (i2cReadBufferIdx >= 5)
-    {
-        if (memcmp(i2cReadBuffer + i2cReadBufferIdx - 5, "<END>", 5) == 0)
-        {
-            end = true;
-        }
-    }
-
-    if(begin && end)
-    {
-        MessageReceiver rcv(i2cReadBuffer + 7, i2cReadBufferIdx - (7 + 5));
-        rcv.processMessage();
-
-        memset(i2cReadBuffer, 0x00, sizeof(i2cReadBuffer));
-        i2cReadBufferIdx = 0;
-    }
-}
-
-I2CInterface::MessageReceiver::MessageReceiver(const uint8_t *msg, size_t msgSize)
-    : _msgSize(msgSize)
-{
-    memcpy(_message, msg, msgSize);
-}
-
-I2CInterface::MessageReceiver::~MessageReceiver()
-{
-}
-
-void I2CInterface::MessageReceiver::processMessage()
-{
-    if (verityCheckSum(_message, _msgSize) == false)
-    {
-        Serial.println("CheckSum error");
-        return;
-    }
-
-    const MsgCommonHeader *msgHeader = (const MsgCommonHeader *)_message;
-    const uint8_t msgId = msgHeader->msgId; // 메시지 ID
-    const uint32_t seqId = ntohl(msgHeader->seqId);
-    const uint8_t numOfData = msgHeader->numOfData; // 데이터의 수
-
-    Serial.print("MsgId:0x");
-    Serial.print(msgId,16);
-    Serial.print(" SeqId:");
-    Serial.println(seqId);
-
-    switch (msgId)
-    {
-    case MsgId::RequestGet:
-    {
-        const RequestGetData *dataArr = (const RequestGetData *)(_message + sizeof(MsgCommonHeader));
-        CallBack::invokeRequestGetCallBack(seqId, dataArr, numOfData);
-    }
-    break;
-    case MsgId::RequestSet:
-    {
-        const RequestSetData *dataArr = (const RequestSetData *)(_message + sizeof(MsgCommonHeader));
-        CallBack::invokeRequestSetCallBack(seqId, dataArr, numOfData);
-    }
-    break;
-    case MsgId::Subscribe:
-    {
-        const SubscribeData *dataArr = (const SubscribeData *)(_message + sizeof(MsgCommonHeader));
-        CallBack::invokeSubscribeCallBack(seqId, dataArr, numOfData);
-    }
-    break;
-    case MsgId::Unsubscribe:
-    {
-        const UnsubscribeData *dataArr = (const UnsubscribeData *)(_message + sizeof(MsgCommonHeader));
-        CallBack::invokeUnsubscribeCallBack(seqId, dataArr, numOfData);
-    }
-    break;
-    case MsgId::Notification:
-    {
-        const NotificationData *dataArr = (const NotificationData *)(_message + sizeof(MsgCommonHeader));
-        CallBack::invokeNotificationCallBack(seqId, dataArr, numOfData);
-    }
-    break;
-    case MsgId::ResponseGet:
-    {
-        const ResponseGetData *dataArr = (const ResponseGetData *)(_message + sizeof(MsgCommonHeader));
-        CallBack::invokeResponseGetCallBack(seqId, dataArr, numOfData);
-    }
-    break;
-    case MsgId::Acknowledge:
-    {
-        CallBack::invokeAcknowledgeCallBack(seqId, msgId);
-    }
-    break;
-    default:
-        return;
-    }
-}
-
-I2CInterface::MessageSender::MessageSender(const unsigned char msgId)
-    : _header((MsgCommonHeader *)_messageBuffer)
-{
-    memset(_messageBuffer, 0x00, sizeof(_messageBuffer));
-    _header->msgId = msgId;
-    _header->msgSize = sizeof(MsgCommonHeader);
-}
-
-I2CInterface::MessageSender::~MessageSender()
-{
-}
-
-void I2CInterface::MessageSender::setSeqId(uint32_t seqId)
-{
-    _header->seqId = htonl(seqId);
-}
-
-void I2CInterface::MessageSender::sendMessage()
-{
-    noInterrupts();
-
-    uint8_t checkSum = getCheckSum(_messageBuffer, _header->msgSize);
-    _messageBuffer[_header->msgSize] = checkSum;
-    _header->msgSize += 1;
-
-    clearWriteBuffer();
-    writeWriteBuffer((uint8_t*)"<BEGIN>", 7);
-    writeWriteBuffer(_messageBuffer, _header->msgSize);
-    writeWriteBuffer((uint8_t*)"<END>", 5);
-
-    // char logStr[64] = {
-    //     0,
-    // };
-    // snprintf(logStr, sizeof(logStr), "[%s:%d] Write Send Message. Size(%d)", __FUNCTION__, __LINE__, _header->msgSize);
-    // Serial.println(logStr);
-
-    interrupts();
-}
-
-void I2CInterface::MessageSender::appendRequestGetDataCommon(unsigned char dataType)
-{
-    RequestGetData data;
-    memset(&data, 0x00, sizeof(RequestGetData));
-
-    data.type = dataType;
-
-    appendData(data);
-}
-
-void I2CInterface::MessageSender::appendRequestGetData(unsigned char dataType)
-{
-    if (_header->msgId != MsgId::RequestGet)
-        return;
-
-    appendRequestGetDataCommon(dataType);
-}
-
-void I2CInterface::MessageSender::appendSubscribeData(unsigned char dataType)
-{
-    if (_header->msgId != MsgId::Subscribe)
-        return;
-
-    appendRequestGetDataCommon(dataType);
-}
-
-void I2CInterface::MessageSender::appendUnsubscribeData(unsigned char dataType)
-{
-    if (_header->msgId != MsgId::Unsubscribe)
-        return;
-
-    appendRequestGetDataCommon(dataType);
-}
-
-void I2CInterface::MessageSender::appendResponseGetDataCommon(unsigned char dataType, uint32_t value)
-{
-    ResponseGetData data;
-    memset(&data, 0x00, sizeof(ResponseGetData));
-
-    data.type = dataType;
-    data.value = htonl(value);
-
-    appendData(data);
-}
-
-void I2CInterface::MessageSender::appendResponseGetData(unsigned char dataType, uint32_t value)
-{
-    if (_header->msgId != MsgId::ResponseGet)
-        return;
-
-    appendResponseGetDataCommon(dataType, value);
-}
-
-void I2CInterface::MessageSender::appendNotificationData(unsigned char dataType, uint32_t value)
-{
-    if (_header->msgId != MsgId::Notification)
-        return;
-
-    appendResponseGetDataCommon(dataType, value);
-}
-
-void I2CInterface::MessageSender::appendRequestSetData(unsigned char dataType, uint32_t value)
-{
-    if (_header->msgId != MsgId::RequestSet)
-        return;
-
-    appendResponseGetDataCommon(dataType, value);
-}
 
 void I2CInterface::I2C_Master::init()
 {
