+++ src/MessageInterface/MessageInterface.h
@@ -1,0 +1,10 @@
+#ifndef __MESSAGE_INTERFACE_H__
+#define __MESSAGE_INTERFACE_H__
+
+#include "MessageInterfaceDefinition.h"
+#include "MessageInterfaceUtil.h"
+#include "MessageCallback.h"
+#include "MessageSender.h"
+#include "MessageReceiver.h"
+
+#endif
